ecommerce/requirements.txt
fastapi 
uvicorn[standard]
sqlalchemy
alembic
pymysql
cryptography
motor
pymongo
pydantic
pydantic-settings
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
httpx
python-dotenv
